Deal with "Delegating orders to arel" deprecation message
This section of code confuses me because it uses a long-winded method
to append the random clause to the orders even though orders are
normally appended anyway. I'm currently testing against AR 5.2.1 but
I'm fairly sure this has always been the case. I do recall a bug in
Squeel that caused orders to be prepended instead but I doubt that was
the reason the code was written this way. In any case, simply
appending in the usual manner deals with the deprecation message.